+++ pynautobot/core/app.py
@@ -92,7 +92,10 @@ class App(object):
          'Testfield2': {'Othervalue2': 4, 'Othervalue1': 3}}
         """
         custom_fields = Request(
-            base="{}/{}/custom-fields/".format(self.api.base_url, self.name,),
+            base="{}/{}/custom-fields/".format(
+                self.api.base_url,
+                self.name,
+            ),
             token=self.api.token,
             http_session=self.api.http_session,
         ).get()
@@ -115,7 +118,10 @@ class App(object):
                                                 'tags']}}}
         """
         config = Request(
-            base="{}/{}/config/".format(self.api.base_url, self.name,),
+            base="{}/{}/config/".format(
+                self.api.base_url,
+                self.name,
+            ),
             token=self.api.token,
             http_session=self.api.http_session,
         ).get()
@@ -153,7 +159,9 @@ class PluginsApp(object):
         }]
         """
         installed_plugins = Request(
-            base="{}/plugins/installed-plugins".format(self.api.base_url,),
+            base="{}/plugins/installed-plugins".format(
+                self.api.base_url,
+            ),
             token=self.api.token,
             http_session=self.api.http_session,
         ).get()