+++ tests/unit/test_endpoint.py
@@ -28,7 +28,11 @@ class EndPointTestCase(unittest.TestCase):
         app = Mock(name="test")
         test_obj = Endpoint(api, app, "test")
         with self.assertRaises(ValueError) as _:
-            test_obj.filter(id=1)
+            test_obj.filter(pk=1)
+        with self.assertRaises(ValueError) as _:
+            test_obj.filter(limit=1)
+        with self.assertRaises(ValueError) as _:
+            test_obj.filter(offset=1)
 
     def test_choices(self):
         with patch("pynautobot.core.query.Request.options", return_value=Mock()) as mock: